iati/requirements.txt
Django==2.0.5
django-modeltranslation==0.13b1
psycopg2-binary==2.7.4
requests==2.18.4
wagtail==2.0.1
wagtail-modeltranslation==0.9.0
